Fix percentage calculation in Student class
Corrects the calcPercentage method to use floating-point division for accurate percentage calculation instead of integer division.

+++ Exceptions_Generics/ExceptionHandlingExamp6/Student.java
@@ -13,7 +13,7 @@ public class Student {
     }
 
     public double calcPercentage(int marks,int total) throws Exception{
-        return (marks/total)*100;
+        return ((double) marks /total)*100;
     }
 
     public String toString(){
